LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-replaceState-expected.txt


FAIL history.replaceState() fires the navigate event assert_equals: expected (object) null but got (undefined) undefined
